LinkedList problem solution - alternateSplit()
Problem statement:-
Write a function AlternatingSplit() which takes one list and divides up
its nodes to make
two smaller lists. The sublists should be made from alternating elements
in the original
list. So if the original list is {a, b, a, b, a}, then one sublist
should be {a, a, a} and the
other should be {b, b}. You may want to use MoveNode() as a helper. The
elements in
the new lists may be in any order (for some implementations, it turns
out to be convenient
if they are in the reverse order from the original list.)